backend/src/services/mcp_server.py
"""
MCP (Model Context Protocol) Server for Todo Task Operations
This module implements the MCP server that exposes todo task operations as stateless tools.
"""

from typing import Dict, Any, Optional
from datetime import datetime, timezone
import logging

# Placeholder for MCP SDK imports until the package is installed
# from python_mcp import Server, tool

logger = logging.getLogger(__name__)


class MCPTaskServer:
    """
    MCP Server implementation for todo task operations.
    Exposes tools for add_task, list_tasks, update_task, complete_task, delete_task.
    """

    def __init__(self):
        self.server = None  # Will hold the MCP server instance once initialized

    def initialize_server(self):
        """
        Initialize the MCP server with all required tools.
        This is a placeholder implementation that will be updated once the MCP SDK is available.
        """
        # Import MCP SDK when available
        try:
            from python_mcp import Server
            self.server = Server(name="todo-task-mcp-server")

            # Register all the required tools
            self._register_add_task_tool()
            self._register_list_tasks_tool()
            self._register_update_task_tool()
            self._register_complete_task_tool()
            self._register_delete_task_tool()

            logger.info("MCP Task Server initialized successfully")
        except ImportError:
            logger.warning("MCP SDK not available. Using mock implementation.")
            # Mock implementation for development
            self.server = MockMCPServer()

    def _register_add_task_tool(self):
        """Register the add_task tool with the MCP server"""
        # This will be implemented with the actual MCP SDK
        pass

    def _register_list_tasks_tool(self):
        """Register the list_tasks tool with the MCP server"""
        # This will be implemented with the actual MCP SDK
        pass

    def _register_update_task_tool(self):
        """Register the update_task tool with the MCP server"""
        # This will be implemented with the actual MCP SDK
        pass

    def _register_complete_task_tool(self):
        """Register the complete_task tool with the MCP server"""
        # This will be implemented with the actual MCP SDK
        pass

    def _register_delete_task_tool(self):
        """Register the delete_task tool with the MCP server"""
        # This will be implemented with the actual MCP SDK
        pass

    def run(self, host: str = "localhost", port: int = 8000):
        """
        Run the MCP server.

        Args:
            host: Host address to bind the server to
            port: Port to run the server on
        """
        if hasattr(self.server, 'serve'):
            self.server.serve(host=host, port=port)
        else:
            logger.info(f"MCP Server mock running on {host}:{port}")


class MockMCPServer:
    """
    Mock MCP Server for development purposes until the real MCP SDK is available.
    """

    def __init__(self):
        self.tools = {}
        self.name = "mock-todo-task-mcp-server"

    def serve(self, host: str, port: int):
        """Mock serve method"""
        print(f"Mock MCP Server '{self.name}' running on {host}:{port}")
        print("Available tools: add_task, list_tasks, update_task, complete_task, delete_task")