backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.20
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
PyJWT==2.9.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt<4.0.0
openai>=1.0.0
alembic>=1.0.0
healthcheck>=1.3.0
prometheus-client>=0.20.0